preventing api keys from being deleted from redis cache fixed api key loading not inserting a record in the redis user_id to tokens map cleaned up redis code

+++ neuroglancer_auth/model/api_key.py
@@ -8,6 +8,10 @@ class APIKey(db.Model):
     user_id = db.Column('user_id', db.Integer, db.ForeignKey("user.id"), unique=True, nullable=False)
     key = db.Column(db.String(32), unique=True, nullable=False)
 
+    @staticmethod
+    def get_by_key(key):
+        return APIKey.query.filter_by(key=key).first()
+
     # load api keys into cache if they don't already exist in redis
     # i.e. new deployment or some redis failure
     @staticmethod
@@ -17,7 +21,7 @@ class APIKey(db.Model):
 
         for api_key in api_keys:
             user = User.get_by_id(api_key.user_id)
-            r.set("token_" + api_key.key, json.dumps(user.create_cache()), nx=True)
+            maybe_insert_token(user.id, api_key.key, json.dumps(user.create_cache()), ex=None, force=True)
     
     @staticmethod
     def generate(user_id):
@@ -31,7 +35,7 @@ class APIKey(db.Model):
 
         user = User.get_by_id(user_id)
         user_json = json.dumps(user.create_cache())
-        token = insert_and_generate_unique_token("userid_" + str(user_id), user_json)
+        token = insert_and_generate_unique_token(user.tokens_key(), user_json)
 
         if not new_entry:
             delete_token(user_id, entry.key)
@@ -45,21 +49,14 @@ class APIKey(db.Model):
 
         return token
 
-def insert_and_generate_unique_token(tokens_key, value, ex=None):
-    token = None
+def maybe_insert_token(user_id, token, value, ex=None, force=False):
+    # nx = Only set the key if it does not already exist
+    not_dupe = r.set("token_" + token, value, nx=True, ex=ex)
 
-    # keep trying to insert a random token into redis until it finds one that is not already in use
-    while True:
-        token = secrets.token_hex(16)
-        # nx = Only set the key if it does not already exist
-        not_dupe = r.set("token_" + token, value, nx=True, ex=ex)
+    if not_dupe or force:
+        r.sadd("userid_" + str(user_id), token)
 
-        if not_dupe:
-            break
-
-    r.sadd(tokens_key, token)
-
-    return token
+    return not_dupe
 
 def delete_token(user_id, token):
     p = r.pipeline()
@@ -72,4 +69,13 @@ def delete_all_tokens_for_user(user_id):
 
     for token_bytes in tokens:
         token = token_bytes.decode('utf-8')
-        delete_token(user_id, token)
+        ttl = r.ttl("token_" + token)
+        if ttl != -1: # don't delete API keys
+            delete_token(user_id, token)
+
+def insert_and_generate_unique_token(user_id, value, ex=None):
+    # keep trying to insert a random token into redis until it finds one that is not already in use
+    while True:
+        token = secrets.token_hex(16)
+        if maybe_insert_token(user_id, token, value, ex):
+            return token
